feat : 리더 확인 hook으로 제작
