+++ src/app/_hook/useLeaderCheck.ts
@@ -1,0 +1,22 @@
+// hooks/useLeaderCheck.ts
+import { useEffect, useState } from "react";
+import { getMyCookie } from "@/app/_providers/action";
+
+export function useLeaderCheck() {
+  const [isLeader, setIsLeader] = useState(false);
+
+  useEffect(() => {
+    async function fetchCookie() {
+      const cookie = await getMyCookie();
+      if (cookie?.name === "JSESSIONID") {
+        const role = localStorage.getItem("role");
+        if (role === "ROLE_LEADER") {
+          setIsLeader(true);
+        }
+      }
+    }
+    fetchCookie();
+  }, []);
+
+  return isLeader;
+}
